LDEV-3323 fix exception on releasing marks in monitor

+++ lams_tool_forum/src/java/org/lamsfoundation/lams/tool/forum/service/ForumService.java
@@ -514,7 +514,7 @@ public class ForumService implements IForumService, ToolContentManager, ToolSess
 		report.setDateMarksReleased(new Date());
 		if (notifyLearnersOnMarkRelease) {
 		    ForumUser user = msg.getCreatedBy();
-		    StringBuilder notificationMessage = notificationMessages.get(user.getUserId());
+		    StringBuilder notificationMessage = notificationMessages.get(user.getUserId().intValue());
 		    if (notificationMessage == null) {
 			notificationMessage = new StringBuilder();
 		    }
